quick fix for gcc / clang compilation
also added a short silly 'makefile' for convenience
