+++ main64.cpp
@@ -8,7 +8,7 @@
 
 #include "rans64.h"
 
-#include <intrin.h>
+#include "utils.h"
 
 // This is just the sample program. All the meat is in rans_byte.h.
 
@@ -44,17 +44,6 @@ static uint8_t* read_file(char const* filename, size_t* out_size)
         *out_size = size;
 
     return buf;
-}
-
-#define WIN32_LEAN_AND_MEAN
-#include <Windows.h>
-
-double timer()
-{
-    LARGE_INTEGER ctr, freq;
-    QueryPerformanceCounter(&ctr);
-    QueryPerformanceFrequency(&freq);
-    return 1.0 * ctr.QuadPart / freq.QuadPart;
 }
 
 // ---- Stats
@@ -183,7 +172,7 @@ int main()
     printf("rANS encode:\n");
     for (int run=0; run < 5; run++) {
         double start_time = timer();
-        uint64_t enc_start_time = __rdtsc();
+        uint64_t enc_start_time = rdtsc();
 
         Rans64State rans;
         Rans64EncInit(&rans);
@@ -196,7 +185,7 @@ int main()
         Rans64EncFlush(&rans, &ptr);
         rans_begin = ptr;
 
-        uint64_t enc_clocks = __rdtsc() - enc_start_time;
+        uint64_t enc_clocks = rdtsc() - enc_start_time;
         double enc_time = timer() - start_time;
         printf("%lld clocks, %.1f clocks/symbol (%5.1fMiB/s)\n", enc_clocks, 1.0 * enc_clocks / in_size, 1.0 * in_size / (enc_time * 1048576.0));
     }
@@ -205,7 +194,7 @@ int main()
     // try rANS decode
     for (int run=0; run < 5; run++) {
         double start_time = timer();
-        uint64_t dec_start_time = __rdtsc();
+        uint64_t dec_start_time = rdtsc();
 
         Rans64State rans;
         uint32_t* ptr = rans_begin;
@@ -217,7 +206,7 @@ int main()
             Rans64DecAdvanceSymbol(&rans, &ptr, &dsyms[s], prob_bits);
         }
 
-        uint64_t dec_clocks = __rdtsc() - dec_start_time;
+        uint64_t dec_clocks = rdtsc() - dec_start_time;
         double dec_time = timer() - start_time;
         printf("%lld clocks, %.1f clocks/symbol (%5.1fMiB/s)\n", dec_clocks, 1.0 * dec_clocks / in_size, 1.0 * in_size / (dec_time * 1048576.0));
     }
@@ -236,7 +225,7 @@ int main()
     printf("\ninterleaved rANS encode:\n");
     for (int run=0; run < 5; run++) {
         double start_time = timer();
-        uint64_t enc_start_time = __rdtsc();
+        uint64_t enc_start_time = rdtsc();
 
         Rans64State rans0, rans1;
         Rans64EncInit(&rans0);
@@ -260,7 +249,7 @@ int main()
         Rans64EncFlush(&rans0, &ptr);
         rans_begin = ptr;
 
-        uint64_t enc_clocks = __rdtsc() - enc_start_time;
+        uint64_t enc_clocks = rdtsc() - enc_start_time;
         double enc_time = timer() - start_time;
         printf("%lld clocks, %.1f clocks/symbol (%5.1fMiB/s)\n", enc_clocks, 1.0 * enc_clocks / in_size, 1.0 * in_size / (enc_time * 1048576.0));
     }
@@ -269,7 +258,7 @@ int main()
     // try interleaved rANS decode
     for (int run=0; run < 5; run++) {
         double start_time = timer();
-        uint64_t dec_start_time = __rdtsc();
+        uint64_t dec_start_time = rdtsc();
 
         Rans64State rans0, rans1;
         uint32_t* ptr = rans_begin;
@@ -294,7 +283,7 @@ int main()
             Rans64DecAdvanceSymbol(&rans0, &ptr, &dsyms[s0], prob_bits);
         }
 
-        uint64_t dec_clocks = __rdtsc() - dec_start_time;
+        uint64_t dec_clocks = rdtsc() - dec_start_time;
         double dec_time = timer() - start_time;
         printf("%lld clocks, %.1f clocks/symbol (%5.1fMB/s)\n", dec_clocks, 1.0 * dec_clocks / in_size, 1.0 * in_size / (dec_time * 1048576.0));
     }
